change distance type in correlation coef. table from integer to decimal

+++ backend/src/main/kotlin/cz/cvut/fit/vwm/collaborativefiltering/data/model/Entities.kt
@@ -4,4 +4,4 @@ data class User(val id: Int, val name: String, val surname: String, val email: S
 data class Song(val id: Int = 0, val mbid: String?, val artist: String, val title: String, val lastFmRank: Int? = null, val url: String? = null)
 data class Recommendation(val id: Int, val userId: Int, val songId: Int, val viewed: Boolean)
 data class Review(val id: Int, val userId: Int, val songId: Int, val value: Int)
-data class CorrelationCoeficient(val id: Int, val userId1: Int, val userId2: Int, val distance: Int, val spearmanCoeficient: Double)
+data class CorrelationCoeficient(val id: Int, val userId1: Int, val userId2: Int, val distance: Double, val spearmanCoeficient: Double)
